Reuse the surrogate-mother PDF's image re-encode fix for the clinic logo
@react-pdf/renderer only reliably embeds baseline JPEG/PNG. The surrogate
mother profile PDF export already works around this by re-encoding
uploaded photos through a canvas before handing them to react-pdf (see
reencodePdfImageDataUrl in smallCard/renderTopBlock.js) - a progressive
JPEG or EXIF-rotated photo would otherwise silently fail to appear.

The Documents Builder clinic logo never got this treatment: its data URL
went straight from Storage into the PDF/preview. Extracted the re-encode
helper into src/utils/pdfImageEncoding.js (both call sites now share it)
and apply it to each clinic logo variant right after it's fetched, so the
same logo file that already works for other exports also renders in the
Documents PDF and its on-page preview.
